requirements.txt
# Minimal Python dependencies for nnids_pipeline
boto3>=1.28.0
pandas>=2.0.0
scikit-learn>=1.3.0
PyYAML>=6.0

# Optional but useful for tests
pytest>=7.4.0
